template: move LogTemplateOptions to common_template_typedefs
